Add files via upload
Pie charts are a bit of a different beast. To create a pie chart, you will be using Matplotlib's subplots() function, which returns a Figure and an Axes object. To see how that works, create a new file named pie_chart_plain.py and put this code in it:

In this code, you create subplots() and then use the pie() method of the Axes object. You pass in a list of numbers as you did before, as well as a list of labels. Then when you run the code, you will see your pie chart:

For this example, you use the explode parameter to tell the pie chart to "explode" or remove a slice from the pie. In this case, you remove the first slice, which corresponds to "Python". You also add a shadow to the pie chart. You can tell your pie chart to rotate a certain number of degrees counter-clockwise by setting the startangle. If you'd like to show the slice percentages, you can use autopct, which will use Python's string interpolation syntax.